Code split svgr away
